+++ pyvela/examples/run_example_emcee.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 
 # %%
-import sys
 from timeit import timeit
 
 import corner
@@ -11,42 +10,35 @@ from matplotlib import pyplot as plt
 from argparse import ArgumentParser
 
 from pyvela import SPNTA, Vela as vl
-from pyvela.priors import parse_custom_prior_file
 
 
 parser = ArgumentParser()
 parser.add_argument("par_file")
 parser.add_argument("tim_file")
-parser.add_argument("-P", "--prior_file")
+parser.add_argument("-P", "--priors", default={})
 args = parser.parse_args()
 
 
-# %%
-prior_dict = (
-    parse_custom_prior_file(args.prior_file) if args.prior_file is not None else {}
-)
-
-# %%
 spnta = SPNTA(
-    args.par_file, args.tim_file, cheat_prior_scale=10, custom_priors=prior_dict
+    args.par_file, args.tim_file, cheat_prior_scale=10, custom_priors=args.priors
 )
 
-# %%
+
 shifts = [
     (spnta.model.param_handler._default_params_tuple.F_.x if pname == "F0" else 0)
     for pname in spnta.param_names
 ]
 
-# %%
+
 maxlike_params_v = np.array([spnta.maxlike_params])
 
-# %%
+
 print(spnta.lnpost_vectorized(maxlike_params_v))
 print(
     timeit("spnta.lnpost_vectorized(maxlike_params_v)", globals=globals(), number=1000)
 )
 
-# %%
+
 nwalkers = spnta.ndim * 5
 p0 = np.array(
     [spnta.prior_transform(cube) for cube in np.random.rand(nwalkers, spnta.ndim)]
@@ -61,11 +53,11 @@ sampler = emcee.EnsembleSampler(
 )
 sampler.run_mcmc(p0, 6000, progress=True, progress_kwargs={"mininterval": 1})
 
-# %%
+
 samples_v_0 = sampler.get_chain(flat=True, discard=2500, thin=40)
 samples_v = spnta.rescale_samples(samples_v_0)
 
-# %%
+
 params_no_plot = [
     "PLREDSIN_",
     "PLREDCOS_",
@@ -80,14 +72,14 @@ param_plot_mask = [
     if all(not par.startswith(pnp) for pnp in params_no_plot)
 ]
 
-# %%
+
 means = (np.mean(samples_v_0, axis=0) + shifts) / spnta.scale_factors
 stds = np.std(samples_v, axis=0)
 for idx, (pname, mean, std) in enumerate(zip(spnta.param_names, means, stds)):
     if idx in param_plot_mask:
-        print(f"{pname}\t\t{mean}\t\t{std}")
+        print(f"{pname}\t\t{mean:.18e}\t\t{std:.4e}")
 
-# %%
+
 # param_labels = [f"\n\n{pname}\n({m[pname].units})\n" for pname in param_names]
 param_labels = [
     f"\n\n{label}\n\n"
@@ -109,7 +101,7 @@ plt.suptitle(spnta.model.pulsar_name)
 plt.tight_layout()
 plt.show()
 
-# %%
+
 params_median = vl.read_params(spnta.model, np.median(samples_v_0, axis=0))
 rv = (
     list(map(vl.value, vl.form_residuals(spnta.model, spnta.toas, params_median)))